Improve functionality to plotting functions
This commit allows for multiple matplotlib figures to be used at once in
the plotting functions. See _main() in pred.py for an example.

In addition we have increased the text size to always be "x-large", so
that labels are large enough to read in the PDF output of the paper.